+++ pennylane_lightning/lightning_base/lightning_base.py
@@ -32,7 +32,6 @@ from pennylane.devices import Device, ExecutionConfig
 from pennylane.devices.modifiers import simulator_tracking, single_tape_support
 from pennylane.exceptions import DeviceError
 from pennylane.measurements import MidMeasureMP, Shots, ShotsLike
-from pennylane.operation import Operator
 from pennylane.tape import QuantumScript, QuantumTape
 from pennylane.typing import Result, ResultBatch, TensorLike
 
@@ -46,17 +45,6 @@ Result_or_ResultBatch = Union[Result, ResultBatch]
 QuantumTapeBatch = Sequence[QuantumTape]
 QuantumTape_or_Batch = Union[QuantumTape, QuantumTapeBatch]
 PostprocessingFn = Callable[[ResultBatch], Result_or_ResultBatch]
-
-
-def base_stopping_condition(op: Operator) -> bool:
-    """A function that determines whether or not an operation is supported by ``lightning.qubit``."""
-
-    if isinstance(op, (qml.ops.op_math.SProd, qml.ops.op_math.Exp)):
-        return True
-
-    if op.name in ("C(SProd)", "C(Exp)"):
-        return True
-    return False
 
 
 @simulator_tracking