+++ pennylane_lightning/lightning_qubit/lightning_qubit.py
@@ -47,7 +47,6 @@ from pennylane_lightning.lightning_base.lightning_base import (
     LightningBase,
     QuantumTape_or_Batch,
     Result_or_ResultBatch,
-    base_stopping_condition,
 )
 
 try:
@@ -77,17 +76,17 @@ _to_matrix_ops = {
 
 def stopping_condition(op: Operator) -> bool:
     """A function that determines whether or not an operation is supported by ``lightning.qubit``."""
-
-    if base_stopping_condition(op):
-        return True
-
-    # As ControlledQubitUnitary == C(QubitUnitrary), it can be removed from `_operations` to keep consistency with `lightning_qubit.toml`
+    # As ControlledQubitUnitary == C(QubitUnitrary),
+    # it can be removed from `_operations` to keep
+    # consistency with `lightning_qubit.toml`
     if isinstance(op, qml.ControlledQubitUnitary):
         return True
     if isinstance(op, qml.PauliRot):
         word = op._hyperparameters["pauli_word"]  # pylint: disable=protected-access
         # decomposes to IsingXX, etc. for n <= 2
         return reduce(lambda x, y: x + (y != "I"), word, 0) > 2
+    if op.name in ("C(SProd)", "C(Exp)"):
+        return True
 
     if (isinstance(op, Conditional) and stopping_condition(op.base)) or isinstance(
         op, MidMeasureMP