+++ .github/CHANGELOG.md
@@ -26,6 +26,7 @@
   PennyLane. See PennyLane PR [(#7944)](https://github.com/PennyLaneAI/pennylane/pull/7944) for
   more information.
   [(#1226)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1226)
+  [(#1246)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1246)
 
 - Dropped support for Python 3.10. Supported Python versions are now 3.11, 3.12 and 3.13.
   [(#1224)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1224)
@@ -45,8 +46,9 @@
 
 <h3>Bug fixes 🐛</h3>
 
-- Update stopping condition for `qml.ops.op_math.SProd` and `qml.ops.op_math.Exp`, in all devices, after breaking changes upstream.
+- **CHANGE REVERTED:** Update stopping condition for `qml.ops.op_math.SProd` and `qml.ops.op_math.Exp`, in all devices, after breaking changes upstream. Was instead fixed in PennyLane and change was reverted.
   [(#1239)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1239)
+  [(#1246)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1246)
 
 - Setup the correct PL version for stable-stable test with `lightning.gpu` with MPI.
   [(#1237)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1237)